+++ packages/adminui-vue/components/navbarSlotDefault/UNavbarSearchButton.vue
@@ -60,7 +60,7 @@
       type="daterange"
       :start-placeholder="$ut('el').datepicker.startDate"
       :end-placeholder="$ut('el').datepicker.endDate"
-      format="dd.MM.yyyy"
+      :format="$ut('el').datepicker.format.date"
       :picker-options="pickerOptions"
     />
   </el-popover>
@@ -88,6 +88,7 @@ export default {
       period: null,
       isPeriod: false,
       pickerOptions: {
+        firstDayOfWeek: this.$ut('el.datepicker.format.firstDayOfWeek'),
         shortcuts: [{
           text: this.$ut('el').datepicker.today,
           onClick (picker) {